+++ comb_spec_searcher/comb_spec_searcher.py
@@ -26,7 +26,7 @@ from logzero import logger
 from comb_spec_searcher.typing import CombinatorialClassType, CSSstrategy
 
 from .class_db import ClassDB
-from .class_queue import DefaultQueue
+from .class_queue import CSSQueue, DefaultQueue
 from .exception import (
     ExceededMaxtimeError,
     InvalidOperationError,
@@ -70,6 +70,8 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
         strategy_pack: StrategyPack,
         *,
         ruledb: Optional[RuleDBAbstract] = None,
+        classdb: Optional[ClassDB[CombinatorialClassType]] = None,
+        classqueue: Optional[CSSQueue] = None,
         expand_verified: bool = False,
         debug: bool = False,
     ):
@@ -95,9 +97,16 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
 
         self.func_times: Dict[str, float] = defaultdict(float)
         self.func_calls: Dict[str, int] = defaultdict(int)
+        self.func_yield: Dict[str, int] = defaultdict(int)
 
-        self.classdb = ClassDB[CombinatorialClassType](type(start_class))
-        self.classqueue = DefaultQueue(strategy_pack)
+        self.classdb = (
+            classdb
+            if classdb is not None
+            else ClassDB[CombinatorialClassType](type(start_class))
+        )
+        self.classqueue = (
+            DefaultQueue(strategy_pack) if classqueue is None else classqueue
+        )
         self.ruledb: RuleDBAbstract = ruledb if ruledb is not None else RuleDB()
         self.ruledb.link_searcher(self)
 
@@ -272,16 +281,14 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
         - add class to classqueue
         """
         for comb_class, child_label in zip(rule.children, end_labels):
+            if not rule.possibly_empty:
+                self.classdb.set_empty(child_label, empty=False)
             if self.symmetries and child_label not in self.symmetry_expanded:
-                self._symmetry_expand(
-                    comb_class, child_label
-                )  # TODO: mark symmetries as empty where appropriate
+                self._symmetry_expand(comb_class, child_label)
             if rule.workable:
                 self.classqueue.add(child_label)
             if not rule.inferrable:
                 self.classqueue.set_not_inferrable(child_label)
-            if not rule.possibly_empty:
-                self.classdb.set_empty(child_label, empty=False)
             self.try_verify(comb_class, child_label)
         if rule.ignore_parent:
             self.classqueue.set_stop_yielding(start_label)
@@ -290,11 +297,13 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
     def _symmetry_expand(self, comb_class: CombinatorialClassType, label: int) -> None:
         """Add symmetries of combinatorial class to the database."""
         sym_labels = set([label])
+        empty = self.classdb.is_empty(comb_class, label)
         for strategy_generator in self.symmetries:
             for start_label, end_labels, rule in self._expand_class_with_strategy(
                 comb_class, strategy_generator, label=label
             ):
                 sym_label = end_labels[0]
+                self.classdb.set_empty(sym_label, empty)
                 self.ruledb.add(start_label, (sym_label,), rule)
                 self.classqueue.set_stop_yielding(sym_label)
                 sym_labels.add(sym_label)
@@ -372,7 +381,7 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
         return status
 
     def _css_status(self, total: float) -> str:
-        table: List[Tuple[str, str, timedelta, str]] = []
+        table: List[Tuple[str, str, timedelta, str, str]] = []
         for explanation in self.func_calls:
             count = f"{self.func_calls[explanation]:,d}"
             time_spent = timedelta(seconds=int(self.func_times[explanation]))
@@ -380,16 +389,26 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
                 percentage = "? %"
             else:
                 percentage = f"{int((self.func_times[explanation] * 100) / total)}%"
-            table.append((explanation, count, time_spent, percentage))
+            if explanation in self.func_yield:
+                yielded = f"{self.func_yield[explanation]:,d}"
+            else:
+                yielded = "-"
+            table.append((explanation, count, time_spent, percentage, yielded))
         table.sort(key=lambda row: row[2], reverse=True)
-        headers = ["", "Number of \napplications", "\nTime spent", "\nPercentage"]
-        colalign = ("left", "right", "right", "right")
+        headers = [
+            "",
+            "Number of\napplications",
+            "\nTime spent",
+            "\nPercentage",
+            "Number of\nrules",
+        ]
+        colalign = ("left", "right", "right", "right", "right")
         return (
             "    "
             + tabulate.tabulate(table, headers=headers, colalign=colalign).replace(
                 "\n", "\n    "
             )
-            + "\n"
+            + "\n\n"
         )
 
     def _mem_status(self, elaborate: bool) -> str: